backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
requests==2.32.3
rapidfuzz==3.9.6
openai==1.51.2
sentence-transformers
numpy
scikit-learn
joblib
groq
httpx==0.27.2